refactor: separate Command module, explicit redirection tokens handling and fix Interpreter::parse_line
