+++ src/cli/run_command.rs
@@ -1,8 +1,8 @@
 use std::process::exit;
 
-use libc::{STDERR_FILENO, STDIN_FILENO, STDOUT_FILENO, c_int, close, dup2, fork, waitpid};
+use libc::{STDERR_FILENO, STDIN_FILENO, STDOUT_FILENO, c_int, close, fork, waitpid};
 
-use crate::{cmd::*, compiler::interpreter::Command};
+use crate::{cmd::*, compiler::command::Command};
 
 // forks a command and returns exit status
 pub fn run_command(cmd: Command) -> i32 {
