+++ src/compiler/mod.rs
@@ -1,3 +1,4 @@
+pub mod command;
 pub mod interpreter;
 pub mod nodes;
 pub mod parser;
